third_party/WebKit/LayoutTests/external/wpt/shadow-dom/event-post-dispatch.html

<!DOCTYPE html>
<title>Shadow DOM: Event dispatch post result</title>
<meta name="author" title="Eriko Kurimoto" href="mailto:elkurin@google.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/shadow-dom.js"></script>
<script src="resources/event-path-test-helpers.js"></script>

<div id="test1">
  <div id="d1">
    <div id="target"></div>
  </div>
</div>

<script>
let n1 = createTestTree(test1);
document.body.appendChild(n1.test1);
test(() => {
  const log = dispatchEventWithEventLog(n1, n1.target, new Event('my-event', { bubbles: true, composed: true }));
  assert_equals(log.event.target, n1.target);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path without ShadowRoots (composed: true).');

test(() => {
  let log = dispatchEventWithEventLog(n1, n1.target, new Event('my-event', { bubbles: true, composed: false }));
  assert_equals(log.event.target, n1.target);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path without ShadowRoots (composed: false).');
</script>

<div id="test2">
  <div id="host">
    <template id="sr" data-mode="open">
      <div id="target"></div>
    </template>
  </div>
</div>

<script>
let n2 = createTestTree(test2);
document.body.appendChild(n2.test2);
test(() => {
  let log = dispatchEventWithEventLog(n2, n2.target, new Event('my-event', { bubbles: true, composed: true }));
  assert_equals(log.event.target, n2.host);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with an open ShadowRoot (composed: true).');

test(() => {
  let log = dispatchEventWithEventLog(n2, n2.target, new Event('my-event', { bubbles: true, composed: false }));
  assert_equals(log.event.target, null);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with an open ShadowRoot (composed: false).');
</script>

<div id="test3">
  <div id="host">
    <template id="sr" data-mode="closed">
      <div id="target"></div>
    </template>
  </div>
</div>

<script>
let n3 = createTestTree(test3);
document.body.appendChild(n3.test3);
test(() => {
  let log = dispatchEventWithEventLog(n3, n3.target, new Event('my-event', { bubbles: true, composed: true }));
  assert_equals(log.event.target, n3.host);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with a closed ShadowRoot (composed: true).');

test(() => {
  let log = dispatchEventWithEventLog(n3, n3.target, new Event('my-event', { bubbles: true, composed: false }));
  assert_equals(log.event.target, null);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with a closed ShadowRoot (composed: false).');
</script>

<div id="test4">
  <div id="host1">
    <template id="sr" data-mode="open">
      <div id="host2">
        <template id="sr" data-mode="open">
          <div id="target"></div>
        </template>
      </div>
    </template>
  </div>
</div>

<script>
let n4 = createTestTree(test4);
document.body.appendChild(n4.test4);
test(() => {
  let log = dispatchEventWithEventLog(n4, n4.target, new Event('my-event', { bubbles: true, composed: true }));
  assert_equals(log.event.target, n4.host1);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with a nested ShadowRoot (composed: true).');

test(() => {
  let log = dispatchEventWithEventLog(n4, n4.target, new Event('my-event', { bubbles: true, composed: false }));
  assert_equals(log.event.target, null);
  assert_equals(log.event.eventPhase, 0);
  assert_equals(log.event.currentTarget, null);
  assert_equals(log.event.composedPath().length, 0);
}, 'Event Path with a nested ShadowRoot (composed: false).');
</script>
